+++ CHANGELOG.md
@@ -12,6 +12,7 @@ _This release is scheduled to be released on 2023-10-01._
 ### Added
 
 - Added UV Index support to OpenWeatherMap
+- Added 'hideDuplicates' flag to the calendar module
 
 ### Removed
 
@@ -26,7 +27,6 @@ _This release is scheduled to be released on 2023-10-01._
 - Fix undefined formatTime method in clock module (#3143)
 - Fix clientonly startup fails after async added (#3151)
 - Fix electron width/heigth when using xrandr under bullseye
-- Fix calendar not showing events with the same name and start date but different calendar url
 
 ## [2.24.0] - 2023-07-01